Extract shared stream-to-buffer utility from ElevenLabs providers
Both ElevenLabsMusic and ElevenLabsSFX contained identical
stream-to-ArrayBuffer conversion code. Extract it into a shared
`streamToBuffer` helper in `lib/providers/stream.ts`, following
the same pattern as the existing `poll.ts` utility.
