+++ lib/providers/music/elevenlabs.ts
@@ -1,6 +1,7 @@
 import { ElevenLabsClient } from "@elevenlabs/elevenlabs-js";
 import type { MusicGenerateParams } from "@/lib/connectors/types";
 import { BaseProvider } from "../base";
+import { streamToBuffer } from "../stream";
 
 export class ElevenLabsMusic extends BaseProvider<
   MusicGenerateParams,
@@ -21,22 +22,6 @@ export class ElevenLabsMusic extends BaseProvider<
       outputFormat: "mp3_22050_32",
     });
 
-    const chunks: Uint8Array[] = [];
-    const reader = stream.getReader();
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      chunks.push(value);
-    }
-
-    const totalLength = chunks.reduce((sum, c) => sum + c.length, 0);
-    const buffer = new Uint8Array(totalLength);
-    let offset = 0;
-    for (const chunk of chunks) {
-      buffer.set(chunk, offset);
-      offset += chunk.length;
-    }
-
-    return buffer.buffer as ArrayBuffer;
+    return streamToBuffer(stream);
   }
 }
